ARM: enable ZBOOT_ROM_TEXT offset for BCM63XX
Broadcom BCM63XX have a quirky bootloader which will refuse loading
arch/arm/boot/compressed/vmlinux at address 0, just enable 0x8000 to
fixup the ELF entry point and make the bootloader happy. This is only
required when using boot from TFTP (which is not supposed to be the case
for production).

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>