net: bcmgenet: remove legacy MDIO bus controller code
Remove the code to register, read and write to/from the GENET MDIO
controller since all of this code is now leveraging the generic UniMAC
MDIO bus controller driver.

Interrupts are now properly handled by the MDIO bus controller as well,
so we keep enabling them, but we do return IRQ_NONE for the interrupt
sharing to work as expected.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>