net: systemport: Add support for switch tag HW extraction
The Broadcom SYSTEMPORT Ethernet controller is capable of extracting a
switch tag (4-bytes Broadcom tag) and put it in the second 32-bits word
of its pre-pended Receive Status Block. When this feature is requested,
make sure that we can satisfy it by turning on receive checksum offload,
and copy the 32-bits words with the tag into skb->cb[] using the
appropriate helper function: dsa_copy_brcm_tag().

We can no longer use napi_gro_receive() because the GRO context would be
using the full skb->cb[] space on 64-bit hosts (which we also run this
driver on). Due to how DSA is structure and since it lacks any kind of
GRO offload, this is not a problem, so we use netif_receive_skb() here
instead.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>