net: dsa: tag_brcm: Support inserting tags in HW
Some Broadcom Ethernet controllers such as SYSTEMPORT are capable of
inserting the 4-bytes Broadcom tag in hardware, in that case, just skip
the software inserting and do a transmission using the master network
device. We need to provide the port mapping to the driver in order for
it to tell the HW how to generate a working Broadcom tag.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>