net: systemport: Add support for switch tag HW insertion
The Broadcom SYSTEMPORT Ethernet controller is capable of inserting a
switch tag (4-bytes Broadcom tag) based on a transmit descriptor flag.
When this feature is requested verify whether the Transmit Status Block
(required for TX checksum offload) is enabled, and if so, allow the
transmit queue selection to be overriden by specifying an appropriate
DEST_MAP, provided from the Broadcom tag helper in net/dsa/tag_brcm.c.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>