net: systemport: Add support for NETIF_F_TSO
Add support for doing software TSO, very much like what the mv643xx_eth
driver. We also need to create a special DMA coherent region for the
driver to DMA from and get the TSO headers, but everything else looks
like normal packet transmission, and just hits a slightly different
transmit path.

Since TSO headers are always DMA mapped, the TX reclaim function is
updated to check the DMA address and see if it belongs in the pool of
TSO headers, and if that's the case, just return another packet
completed for TX queue flow control.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>