mfd: micon: Add Buffalo Kurobox Pro, Terastation II Pro/Live driver
This driver is currently only used to reboot the devices, but the
microcontroller hanging off UART1 on the Buffalo Kurobox Pro and
Terastation II Pro/Live is capable of a lot more than that, which we
will support in subsequent patches. For now, just add the reboot
functionality to make the kernel reboot correctly.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>