net: systemport: Turn features on by default
The most sensible features we have: RXCSUM, IP_CSUM and IPV6_CSUM should
really be turned on by default, so make this happen. We slightly rework
bcm_sysport_set_features() to be less smart and just apply the features
that are requested. This allows us to call this function during
interface open and make resume and make sure that everything is
configured correctly at the hardware level.

Finally, propagate dev->features to dev->vlan_features in order to get
the appropriate stacked devices enabled with the same feature set.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>