net: macb: Remove CONFIG_OF around DT match table
A subsequent patch is going to make of_match_node() an inline stub when
CONFIG_OF is disabled, which will let the compiler eliminate unused variables.
In order not to clutter the code more, remove the CONFIG_OF #ifdef such that
macb_dt_ids and what it references are always defined.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>