net: dsa: debugfs: add port vlan
Add a debug filesystem "vlan" entry to query a port's hardware VLAN
entries through the .port_vlan_dump switch operation.

This is really convenient to query directly the hardware or inspect DSA
or CPU links, since these ports are not exposed to userspace.

Here are the VLAN entries for a CPU port:

    # cat port5/vlan
    vid 1
    vid 42  pvid

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>