net: phy: Export phy_stop_machine()
phy_stop_machine() is publicly exported in include/linux/phy.y, and is
not made static because it's also used by phy_device.c. Since
phy_start_machine() is already exported, do this here too. This is a
function that provides hard guarantees that the state machine is
properly stopped past that synchronization point.

Fixes: 00db8189d984 ("This patch adds a PHY Abstraction Layer to the Linux Kernel, enabling ethernet drivers to remain as ignorant as is reasonable of the connected PHY's design and operation details.")
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>