net: smsc911x: Do not update stats when not running
We cannot perform register accesses in smsc911x_tx_update_txcounters()
because that could happen when the interface is not running and is
powered down. Accessing registers during that context could lead to all
kinds of aborts depending on how the register bus is wired back to the
CPU doing these accesses.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>